requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fsspec==2024.3.1
grpcio==1.62.1
grpcio-tools==1.62.1
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter-client==8.6.1
jupyter-core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tornado==6.4
traitlets==5.14.2
typing-extensions==4.10.0
wcwidth==0.2.13
zipp==3.18.1
